requirements.txt
altgraph==0.17.5
et_xmlfile==2.0.0
lxml==6.0.2
macholib==1.16.4
numpy==2.3.5
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.17.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
setuptools==80.9.0
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
